requirements.txt
Async-Keep-Alive-http-client==develop
PyYAML==3.10
boto==2.17.0
botocore==0.20.0
coverage==3.7
ipdb==0.8
ipython==1.1.0
jmespath==0.1.0
nose==1.3.0
ply==3.4
pyaml==13.07.1
python-dateutil==2.2
requests==2.0.0
shortuuid==0.4
six==1.4.1
tornado==3.1.1
virtualenv==1.10.1
wsgiref==0.1.2
xmltodict==0.8.3
